requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
